+++ sdk/identity/azure-identity/src/client_secret_credential.cpp
@@ -3,6 +3,7 @@
 
 #include "azure/identity/client_secret_credential.hpp"
 
+#include "private/tenant_id_resolver.hpp"
 #include "private/token_credential_impl.hpp"
 
 using Azure::Identity::ClientSecretCredential;
@@ -13,6 +14,7 @@ using Azure::Core::Credentials::AccessToken;
 using Azure::Core::Credentials::TokenCredentialOptions;
 using Azure::Core::Credentials::TokenRequestContext;
 using Azure::Core::Http::HttpMethod;
+using Azure::Identity::_detail::TenantIdResolver;
 using Azure::Identity::_detail::TokenCredentialImpl;
 
 ClientSecretCredential::ClientSecretCredential(
@@ -20,8 +22,10 @@ ClientSecretCredential::ClientSecretCredential(
     std::string const& clientId,
     std::string const& clientSecret,
     std::string const& authorityHost,
+    std::vector<std::string> additionallyAllowedTenants,
     TokenCredentialOptions const& options)
-    : TokenCredential("ClientSecretCredential"), m_clientCredentialCore(tenantId, authorityHost),
+    : TokenCredential("ClientSecretCredential"),
+      m_clientCredentialCore(tenantId, authorityHost, additionallyAllowedTenants),
       m_tokenCredentialImpl(std::make_unique<TokenCredentialImpl>(options)),
       m_requestBody(
           std::string("grant_type=client_credentials&client_id=") + Url::Encode(clientId)
@@ -34,7 +38,13 @@ ClientSecretCredential::ClientSecretCredential(
     std::string const& clientId,
     std::string const& clientSecret,
     ClientSecretCredentialOptions const& options)
-    : ClientSecretCredential(tenantId, clientId, clientSecret, options.AuthorityHost, options)
+    : ClientSecretCredential(
+        tenantId,
+        clientId,
+        clientSecret,
+        options.AuthorityHost,
+        options.AdditionallyAllowedTenants,
+        options)
 {
 }
 
@@ -48,6 +58,7 @@ ClientSecretCredential::ClientSecretCredential(
         clientId,
         clientSecret,
         ClientSecretCredentialOptions{}.AuthorityHost,
+        ClientSecretCredentialOptions{}.AdditionallyAllowedTenants,
         options)
 {
 }
@@ -58,13 +69,20 @@ AccessToken ClientSecretCredential::GetToken(
     TokenRequestContext const& tokenRequestContext,
     Context const& context) const
 {
-  auto const scopesStr = m_clientCredentialCore.GetScopesString(tokenRequestContext.Scopes);
+
+  auto const tenantId = TenantIdResolver::Resolve(
+      m_clientCredentialCore.GetTenantId(),
+      tokenRequestContext,
+      m_clientCredentialCore.GetAdditionallyAllowedTenants());
+
+  auto const scopesStr
+      = m_clientCredentialCore.GetScopesString(tenantId, tokenRequestContext.Scopes);
 
   // TokenCache::GetToken() and m_tokenCredentialImpl->GetToken() can only use the lambda argument
   // when they are being executed. They are not supposed to keep a reference to lambda argument to
   // call it later. Therefore, any capture made here will outlive the possible time frame when the
   // lambda might get called.
-  return m_tokenCache.GetToken(scopesStr, tokenRequestContext.MinimumExpiration, [&]() {
+  return m_tokenCache.GetToken(scopesStr, tenantId, tokenRequestContext.MinimumExpiration, [&]() {
     return m_tokenCredentialImpl->GetToken(context, [&]() {
       auto body = m_requestBody;
 
@@ -73,7 +91,7 @@ AccessToken ClientSecretCredential::GetToken(
         body += "&scope=" + scopesStr;
       }
 
-      auto const requestUrl = m_clientCredentialCore.GetRequestUrl();
+      auto const requestUrl = m_clientCredentialCore.GetRequestUrl(tenantId);
 
       auto request
           = std::make_unique<TokenCredentialImpl::TokenRequest>(HttpMethod::Post, requestUrl, body);