+++ sdk/keyvault/azure-security-keyvault-certificates/src/certificate_client.cpp
@@ -3,6 +3,7 @@
 
 #include "azure/keyvault/certificates/certificate_client.hpp"
 
+#include "azure/keyvault/shared/keyvault_challenge_based_authentication_policy.hpp"
 #include "azure/keyvault/shared/keyvault_shared.hpp"
 #include "private/certificate_constants.hpp"
 #include "private/certificate_serializers.hpp"
@@ -76,7 +77,8 @@ CertificateClient::CertificateClient(
     tokenContext.Scopes = {_internal::UrlScope::GetScopeFromUrl(m_vaultUrl)};
 
     perRetrypolicies.emplace_back(
-        std::make_unique<BearerTokenAuthenticationPolicy>(credential, std::move(tokenContext)));
+        std::make_unique<_internal::KeyVaultChallengeBasedAuthenticationPolicy>(
+            credential, std::move(tokenContext)));
   }
   std::vector<std::unique_ptr<HttpPolicy>> perCallpolicies;